# Conflicts:
#	README.md
#	bin/main/db/migration/mdmc_db.sql
#	build.gradle
#	src/main/java/com/mdmc/posofmyheart/api/controllers/OrderController.java
#	src/main/java/com/mdmc/posofmyheart/api/controllers/ProductController.java
#	src/main/java/com/mdmc/posofmyheart/api/exceptions/ResourceNotFoundException.java
#	src/main/java/com/mdmc/posofmyheart/application/configurations/CorsConfig.java
#	src/main/java/com/mdmc/posofmyheart/application/dtos/OrderResponse.java
#	src/main/java/com/mdmc/posofmyheart/application/mappers/OrderMapper.java
#	src/main/java/com/mdmc/posofmyheart/application/mappers/ProductExtraMapper.java
#	src/main/java/com/mdmc/posofmyheart/application/mappers/ProductMapper.java
#	src/main/java/com/mdmc/posofmyheart/application/services/OrderService.java
#	src/main/java/com/mdmc/posofmyheart/application/services/ProductService.java
#	src/main/java/com/mdmc/posofmyheart/application/services/impl/OrderServiceImpl.java
#	src/main/java/com/mdmc/posofmyheart/application/services/impl/ProductServiceImpl.java
#	src/main/java/com/mdmc/posofmyheart/domain/dtos/CategoryResponseDTO.java
#	src/main/java/com/mdmc/posofmyheart/domain/dtos/CategoryWrapperDTO.java
#	src/main/java/com/mdmc/posofmyheart/domain/dtos/CreateOrderResponse.java
#	src/main/java/com/mdmc/posofmyheart/domain/dtos/ProductOptionDTO.java
#	src/main/java/com/mdmc/posofmyheart/domain/dtos/ProductsWithExtrasDto.java
#	src/main/java/com/mdmc/posofmyheart/domain/models/Category.java
#	src/main/java/com/mdmc/posofmyheart/domain/models/Product.java
#	src/main/java/com/mdmc/posofmyheart/domain/models/ProductExtra.java
#	src/main/java/com/mdmc/posofmyheart/domain/models/ProductVariant.java
#	src/main/java/com/mdmc/posofmyheart/infrastructure/persistence/entities/OrderDetailEntity.java
#	src/main/java/com/mdmc/posofmyheart/infrastructure/persistence/entities/OrderEntity.java
#	src/main/java/com/mdmc/posofmyheart/infrastructure/persistence/entities/PaymentMethodEntity.java
#	src/main/java/com/mdmc/posofmyheart/infrastructure/persistence/entities/ProductCategoryEntity.java
#	src/main/java/com/mdmc/posofmyheart/infrastructure/persistence/entities/ProductEntity.java
#	src/main/java/com/mdmc/posofmyheart/infrastructure/persistence/entities/ProductExtraEntity.java
#	src/main/java/com/mdmc/posofmyheart/infrastructure/persistence/entities/ProductVariantEntity.java
#	src/main/java/com/mdmc/posofmyheart/infrastructure/persistence/entities/SauceEntity.java
#	src/main/java/com/mdmc/posofmyheart/infrastructure/persistence/repositories/OrderDetailRepository.java
#	src/main/java/com/mdmc/posofmyheart/infrastructure/persistence/repositories/PaymentMethodRepository.java
#	src/main/java/com/mdmc/posofmyheart/infrastructure/persistence/repositories/ProductCategoryRepository.java
#	src/main/resources/application-dev.properties
#	src/main/resources/application-remote.properties
#	src/main/resources/db/migration/mdmc.md

+++ src/main/java/com/mdmc/posofmyheart/domain/dtos/CreateOrderResponse.java
@@ -1,5 +1,5 @@
 package com.mdmc.posofmyheart.domain.dtos;
 
 public record CreateOrderResponse(
-        Integer idOrder
+        Long idOrder
 ) {}
